% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{Annotates peaks with genes in gene window}
\usage{
annotate_peaks(
  gene_metadata.dt = gene_metadata.dt,
  atac.sce = atac.sce,
  distance = 150000
)
}
\arguments{
\item{gene_metadata.dt}{a data table of gene metadata}

\item{atac.sce}{a single-cell experiment object}

\item{distance}{a numeric value specifying the distance from the gene body to the peak}
}
\value{
a data table with the overlap between peaks and genes, including the distance between the peak and the gene body
}
\description{
Annotates peaks with genes in gene window
}
