% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_TF_acc.R
\name{cor_TF_acc}
\alias{cor_TF_acc}
\title{Correlate TF expression and Region accessibility}
\usage{
cor_TF_acc(
  rna.sce,
  atac.sce,
  TFs_filt = NULL,
  motifmatcher.se = motifmatcher.se,
  motif2gene.dt = motif2gene.dt,
  correlation_method = "pearson",
  remove_motifs = c("T_789"),
  cores = detectCores()
)
}
\arguments{
\item{rna.sce}{A SingleCellExperiment object containing RNA-seq data.}

\item{atac.sce}{A SingleCellExperiment object containing ATAC-seq data.}

\item{TFs_filt}{Character vector containing the names of TFs to include in the analysis. Default is NULL.}

\item{motifmatcher.se}{SingleCellExperiment object containing motif matches.}

\item{motif2gene.dt}{Data table containing the mapping of motifs to genes.}

\item{correlation_method}{Method to calculate correlation. Default is "pearson", other options include "spearman" and "kendall".}

\item{remove_motifs}{character vector of motifs to be manually removed}

\item{cores}{Number of cores to use for parallel processing.}
}
\value{
A SummarizedExperiment object containing the correlation matrix and the p-value matrix.
}
\description{
This function calculates the correlation between the expression of transcription factors (TFs) and the accessibility of their binding sites across different genomic regions.
}
