% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{matrix.please}
\alias{matrix.please}
\title{Turn data table to matrix}
\usage{
matrix.please(x)
}
\arguments{
\item{x}{a data table to be converted to a matrix.}
}
\value{
a matrix with row names taken from the first column of the input data table.
}
\description{
This function converts a data table (excluding the first column) to a matrix and sets the row names to the values of the first column.
}
