% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{minmax.normalisation}
\alias{minmax.normalisation}
\title{Minmax normalization}
\usage{
minmax.normalisation(x)
}
\arguments{
\item{x}{a numeric vector to be normalized.}
}
\value{
a numeric vector with values between 0 and 1.
}
\description{
This function normalizes a numeric vector to the range [0,1] using min-max normalization.
}
