% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{sort.abs}
\alias{sort.abs}
\title{Sort absolute}
\usage{
\method{sort}{abs}(dt, sort.field)
}
\arguments{
\item{dt}{a data table to be sorted.}

\item{sort.field}{a character string indicating the name of the column to sort based on its absolute value.}
}
\value{
a sorted data table.
}
\description{
This function sorts a data table based on the absolute values of a specified column.
}
