% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF2Gene.R
\name{TF2Gene}
\alias{TF2Gene}
\title{TF2Gene function: Linking TFs to target genes}
\usage{
TF2Gene(
  rna.sce = rna.sce,
  motifmatcher_chip.se = motifmatcher_chip.se,
  peak2gene.dt = peak2gene.dt,
  gene_metadata.dt = NULL,
  atac.sce = NULL,
  filter_genes = TRUE,
  min_chip_score = 0.15,
  distance = 50000,
  cores = detectCores()
)
}
\arguments{
\item{rna.sce}{A SingleCellExperiment object containing RNA-seq data.}

\item{motifmatcher_chip.se}{A MotifMatchR object (output of silico_chip function())}

\item{peak2gene.dt}{Data table linking genomic peaks with genes}

\item{gene_metadata.dt}{(optional, Only needed to find peak-genes if peak2gene.dt not supplied) gene metadata used to annotate peaks with genes}

\item{atac.sce}{(optional, Only needed to find peak-genes if peak2gene.dt not supplied) A SingleCellExperiment object containing the ATAC-seq data}

\item{filter_genes}{logical indicating whether to filter out non-informative genes (default: TRUE)}

\item{min_chip_score}{minimum ChIP score for a peak to be considered a target of a TF (default: 0.15)}

\item{distance}{maximum distance (in bp) between a peak and a gene for them to be considered linked (default: 5e4, can not be larger than maximum distance in peak2gene.dt)}

\item{cores}{number of cores to use for parallel processing (default: detectCores())}
}
\value{
a data frame of regression coefficients (beta), p-values, and TFs and genes
}
\description{
This function takes as input an RNA-Seq dataset, a motif-matching scores dataset, and a dataset linking
genomic peaks with genes. It performs a regression analysis to determine the influence of each transcription
factor on each gene. The output is a data frame of regression coefficients (beta), p-values, and TFs and genes.
}
